Test: move endpoint related integration tests under endpoints folder & use shared setup function.
