+++ tests/endpoints/mod.rs
@@ -1,0 +1,4 @@
+mod alliance;
+mod character;
+mod corporation;
+mod util;
